refactor: Remove unused crates and clean up bindings
- Remove unused specado-wasm and specado-providers crates
- Clean up Node.js and Python bindings after worktree merge
- Fix compilation errors in both bindings
- Update error mappings to match current core Error enum
- Fix type compatibility issues (u64 to f64 for NAPI)
- Add proper PyO3 exception handling
- Remove obsolete documentation files
- Add BINDINGS_STATUS.md to track binding health

Both bindings now compile successfully with Node.js fully operational
and Python requiring only standard PyO3 build environment setup.

+++ crates/specado-nodejs/src/types.rs
@@ -100,7 +100,7 @@ pub struct SamplingParams {
     /// Stop sequences
     pub stop: Option<Vec<String>>,
     /// Random seed
-    pub seed: Option<u64>,
+    pub seed: Option<f64>,
 }
 
 /// Token and output limits
@@ -124,7 +124,7 @@ pub struct MediaConfig {
     /// Supported output media types
     pub output_types: Option<Vec<String>>,
     /// Maximum file size in bytes
-    pub max_file_size: Option<u64>,
+    pub max_file_size: Option<f64>,
 }
 
 /// Tool call in a message
@@ -422,5 +422,5 @@ pub struct ResponseMetadata {
     /// Request ID for tracking
     pub request_id: Option<String>,
     /// Response time in milliseconds
-    pub response_time_ms: u64,
+    pub response_time_ms: f64,
 }
